refactor: validator 폴더 생성 및 유효성 검사 코드 이동
