+++ src/validator/PriceValidator.js
@@ -1,5 +1,5 @@
-import handleError from './utils/handleError.js';
-import ERROR_MESSAGES from './constants/errorMessages.js';
+import handleError from '../utils/handleError.js';
+import ERROR_MESSAGES from '../constants/errorMessages.js';
 
 class PriceValidator {
   static isEmptyPrice(inputPrice) {
